wip: (broken) backend support for timer pause
This ought to do for the backend, except I don't think I can make it
work on the frontend with the current setup. Going to come back to
this later, but switching from threading.Timer to Twisted's callLater
is still an improvement so we want to go ahead and keep this.

...except it doesn't work